Added workaround for .Net Native metadata issue caused by WinMDExp. This was causing issues when submitting apps to the store.

+++ Lib/Microsoft.IoT.DeviceHelpers/Input/PushButtonHelper.cs
@@ -210,7 +210,7 @@ namespace Microsoft.IoT.DeviceHelpers.Input
         /// The <see cref="GpioPinValue"/> that indicates the button is pressed. 
         /// The default is <see cref="GpioPinValue.Low"/>.
         /// </value>
-        [DefaultValue(GpioPinValue.Low)]
+        [DefaultValue((int) GpioPinValue.Low)] // HACK: int cast is a workaround for WinMDExp issue - JB 2016/02/19
         public GpioPinValue PressedValue { get { return pressedValue; } set { pressedValue = value; } }
 
         /// <summary>
