+++ Lib/Microsoft.IoT.Devices/Input/Switch.cs
@@ -182,7 +182,7 @@ namespace Microsoft.IoT.Devices.Input
         /// The <see cref="GpioPinValue"/> that indicates the switch is on. 
         /// The default is <see cref="GpioPinValue.Low"/>.
         /// </value>
-        [DefaultValue(GpioPinValue.Low)]
+        [DefaultValue((int) GpioPinValue.Low)] // HACK: int cast is a workaround for WinMDExp issue - JB 2016/02/19
         public GpioPinValue OnValue { get { return onValue; } set { onValue = value; } }
 
         /// <summary>
